requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0

# Database
sqlalchemy[asyncio]>=2.0.35
aiosqlite>=0.20.0
alembic>=1.13.0

# Templates
jinja2>=3.1.4
aiofiles>=24.1.0

# Auth
bcrypt>=4.2.0
itsdangerous>=2.2.0
python-multipart>=0.0.12

# Media
pillow>=11.0.0
python-magic-bin>=0.4.14

# Settings
pydantic-settings>=2.5.0

# Utilities
python-slugify>=8.0.0
markdown>=3.7

# Testing
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Monitoring (optional)
sentry-sdk[fastapi]>=2.14.0
